Add files via upload
I am very dedicated and motivated to grow in field of data analysis and IT field. My passion for mathematics and finances, both of which were part of my engineering degree, supports my strong analytical thinking. Combine with solid communication skills, I bring many of the qualities you are looking for .